add revoked test and add generic validationservice
