+++ decentralizediso20022/src/main/java/com/centiglobe/decentralizediso20022/application/internal/IntValidationService.java
@@ -1,15 +1,13 @@
 package com.centiglobe.decentralizediso20022.application.internal;
 
-import java.io.File;
-import java.security.KeyStore;
 import java.security.cert.X509Certificate;
 
-import javax.annotation.PostConstruct;
-
+import com.centiglobe.decentralizediso20022.application.ValidationService;
 import com.prowidesoftware.swift.model.mx.AbstractMX;
 import com.prowidesoftware.swift.model.mx.BusinessAppHdrV02;
 
 import org.cryptacular.util.CertUtil;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Service;
@@ -24,33 +22,14 @@ import org.springframework.stereotype.Service;
 @Service
 public class IntValidationService {
 
-    @Value("${server.ssl.key-store}")
-    private String KEY_STORE;
-
-    @Value("${server.ssl.key-store-password}")
-    private String KEY_PASS;
-
-    @Value("${server.ssl.keyAlias}")
-    private String KEY_ALIAS;
-
     @Value("${message.bad-from-header}")
     private String BAD_FROM_HEADER;
     
     @Value("${message.bad-to-header}")
     private String BAD_TO_HEADER;
 
-    private X509Certificate us;
-    
-    /**
-     * Initalizes the validation service with the local certificate in the keystore
-     * 
-     * @throws Exception if the certificate could not be obtained
-     */
-    @PostConstruct
-    public void initCertificate() throws Exception {
-        KeyStore keyStore = KeyStore.getInstance(new File(KEY_STORE), KEY_PASS.toCharArray());
-        us = (X509Certificate) keyStore.getCertificate(KEY_ALIAS);
-    }
+    @Autowired
+    private ValidationService vs;
 
     /**
      * Validates the ISO 20022 message. May optionally validate the message against
@@ -77,7 +56,7 @@ public class IntValidationService {
      */
     public void validateHeader(BusinessAppHdrV02 header, X509Certificate cert) throws IllegalArgumentException, NullPointerException {
         String fromDomain = header.getFr().getFIId().getFinInstnId().getNm();
-        String ourCommonName = CertUtil.subjectCN(us);
+        String ourCommonName = CertUtil.subjectCN(vs.getOurCertificate());
 
         if (!fromDomain.equals(ourCommonName))
             throw new IllegalArgumentException(String.format(BAD_FROM_HEADER, fromDomain, ourCommonName));
