+++ decentralizediso20022/src/main/resources/application-test.properties
@@ -1,10 +1,10 @@
 # keystore used in integration tests
-server.ssl.key-store=keystore.p12
+server.ssl.key-store=src/main/resources/keystore.p12
 server.ssl.key-store-password=qwerty
 server.ssl.keyAlias=localhost
 
 # truststore used in integration tests
-server.ssl.test-trust-store=truststore.p12
+server.ssl.test-trust-store=src/main/resources/truststore.p12
 server.ssl.test-trust-store-password=ytrewq
 
 # truststore used for validation
